codedna/requirements.txt
# Endee Vector Database SDK
endee>=0.1.0

# Code Parsing
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-java>=0.21.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Sparse Vectors
scikit-learn>=1.3.0
scipy>=1.11.0

# Frontend
streamlit>=1.28.0
plotly>=5.18.0

# Visualization
matplotlib>=3.8.0
networkx>=3.2.0
umap-learn>=0.5.4

# Utilities
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
gitpython>=3.1.0
rich>=13.7.0
click>=8.1.0
tqdm>=4.66.0
